Extract new-item-modal region into sections/new-item-modal.js
Fifth of P-08's six region extractions. 140 LOC — the largest of the P-08 batch. The form-wiring IIFE becomes initNewItemModalSideEffects(); submitNewItem's post-create tab-switch imports switchTab/switchSubtab/refreshAll from dashboard-main.js (now exported). saveConfig stays put — it was only in this region by accident of textual layout, it belongs to the in-app config editor (region preceding this one) and will move with that region in a later PR.
